Moved functions of the area to DONOTAREA.c
Moved printArea and getAreaSize to DONOTarea.c (it is called DONOTarea.c because of the fact that it should not be compiled separately). Updated README.md
