+++ README.md
@@ -6,6 +6,7 @@ Personal project over the holidays of 2020, during/after quarantine.
 ---
 Written in C using Xcode version 11.5 on macOS Catalina (10.15.5), build version 11E608c (`xcodebuild -version`). 
 When opened in Xcode, automatically runs in Terminal where it is supposed to for compatibility reasons using utf-8 as default encoding, LF as default line breaks. **Xcode shows an invalid conversion in `rlutil.h`, do not attempt to fix it - it will break stuff.**
+Files with the prefix of DONOT should be disabled for separate compilation, they should run only when included in main.c
 
 ---
 Uses `#pragma mark` for visual assistance on the sliding menu in Xcode, may be an issue when running on Windows - `#pragma mark` is used everywhere **EXCEPT**  `rlutil/h`, which is a library by Tapio to have colors in Terminal/cmd.
