+++ snake/main.c
@@ -46,6 +46,8 @@
 #define END             8191    //end of line. if value in end searcher is 8192, it realises that it is the end of the line/coloumn
 
 
+#include "DONOTarea.c"
+
 //  a = area
 //  a[coloumn][row]
 //  program uses [HEIGHT][WIDTH] because double arrays when entered in (every array on 2nd level has a new line) means that the second must be rows and the first variable must be columns
@@ -66,43 +68,7 @@ int area[HEIGHT][WIDTH] = {
 };
 
 //  prints out the area and needs the WIDTH parameter because in memory, the numbers are one behind another so it can see how far it needs to go for a new line
-#pragma mark printArea() function
-void printArea(int a[][WIDTH], int w, int h) {
-    saveDefaultColor(); //saves color that has been used before to set it back later
-    int c;  //current point going through
-    for (int i = 0; i < h; i++) {
-        for (int j = 0; j < w; j++) {
-            c = a[i][j];
-            if (c == WALL) {
-                setColor(LIGHTCYAN);
-                setBackgroundColor(RED);
-                printf("# ");
-            }
-            else if (c == FREE) {
-                setColor(LIGHTGREEN);
-                setBackgroundColor(BLUE);
-                printf("  ");
-            }
-            else if (c == HEAD) {
-                setColor(LIGHTRED);
-                setBackgroundColor(MAGENTA);
-                printf("~ ");
-            }
-            else if (c == DOOR) {
-                setColor(LIGHTMAGENTA);
-                setBackgroundColor(DARKGREY);
-                printf("\\ ");
-            }
-            else {
-                setColor(WHITE);
-                setBackgroundColor(BLACK);
-                printf("? ");
-            }
-        }
-        printf("\n");
-    }
-    resetColor();   //resets color to the colors saved on the start of printArea()
-}
+
 #pragma mark testing()
 void testing() {    //just to test features before putting them into run(), where they are in the 'temporary final' state
     cls();          //clears screen, useful to 'preset' the cmd window
@@ -113,6 +79,10 @@ void testing() {    //just to test features before putting them into run(), wher
 void run() {
     cls();          //clears screen, useful to 'preset' the cmd window
     hidecursor();   //hides the cursor from the terminal window (only the pointer to where you're typing
+    
+    struct areaSize size;
+    size = getAreaSize(area);
+    printArea(area, size.w, size.h);
     return;
 }
 #pragma mark main()
@@ -120,7 +90,7 @@ int main(int argc, const char * argv[]) {
     saveDefaultColor();
     testing();
     
-    //run();
-    resetColor();
+    run();
+    //resetColor();
     return 0;
 }
